updated adatpter, moved adapter code  inline, cleaned up formating issue
