+++ static/v1.2/intent-b/index.html
